Refactor injector testers into separate modules
