+++ src/test/scala/chiffreTests/injector/CycleInjectorSpec.scala
@@ -13,13 +13,21 @@
 // limitations under the License.
 package chiffreTests.inject
 
+import chiffre.inject.CycleInjectorInfo
 import chisel3.iotesters.ChiselFlatSpec
 
 class CycleInjectorSpec extends ChiselFlatSpec {
   behavior of "CycleInjectorInfo"
 
-  it should "generate a sensible name" in (pending)
-  it should "be the expected width" in (pending)
+  it should "generate a sensible name" in {
+    val x = CycleInjectorInfo(512, 1024)
+    x.name should be (s"cycle1024")
+  }
+
+  it should "be the expected width" in {
+    val x = CycleInjectorInfo(2048, 4096)
+    x.width should be (2048 + 4096)
+  }
 
   behavior of "CycleInjector"
